requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
packaging==21.3
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
pkg_resources==0.0.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
requests==2.26.0
responses==0.16.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
SDK==1.0.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
